+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [0.5.2] - 2024-01-06
-- fix #35 add **bool isReady()**
+- fix #35, add **bool isConversionReady()**
+  - thanks to Henk Holdijk
+- add **bool waitConversionReady(uint32_t timeout = INA226_MAX_WAIT_MS)**
+- INA226_MAX_WAIT_MS
   - thanks to Henk Holdijk